refactor: (#11) 폴더 구조 일괄 변경
