+++ src/features/main/ui/MainHeader/SearchBar.tsx
