+++ src/features/main/ui/RoomList/RoomCard.tsx
