--- src/features/main/ui/Rooms/index.ts
@@ -1,8 +1,0 @@
-export { MainHeader } from './MainHeader';
-export { NotificationButton } from './NotificationButton';
-export { RoomCard } from './RoomCard';
-export { RoomFilterTabs } from './RoomFilterTabs';
-export { RoomList } from './RoomList';
-export { SearchBar } from './SearchBar';
-export { SideBar } from './SideBar';
-export { CreateRoomButton } from './CreateRoomButton';
