+++ src/features/main/ui/index.ts
@@ -1,3 +1,5 @@
-export * from './Rooms';
+export * from './RoomList';
 export * from './Chat';
+export * from './MainHeader';
 export * from './CreateRoomModal';
+export * from './SideBar';
